fix(atom): avoid 32-bit shift overflow when computing pool sizes
1U << log2blkpoolsz shifted in 32-bit width even though log2blkpoolsz
sizes a size_t; cast to size_t before shifting so 64-bit pool sizes
work correctly instead of relying on implicit widening after the fact.

+++ libraries/datastruct/atom/new.c
@@ -96,7 +96,7 @@ result_t atom_ensure_blk_space(atom_set_t *s, size_t length)
    * pool, then there's space */
 
   if (s->b_used > 0 &&
-      s->blkpools[s->b_used - 1].used + length <= (1U << s->log2blkpoolsz))
+      s->blkpools[s->b_used - 1].used + length <= ((size_t) 1 << s->log2blkpoolsz))
     return result_OK;
 
   /* otherwise the last allocated pool is full or there are no pools
@@ -209,7 +209,7 @@ result_t atom_new(atom_set_t          *s,
    * case.
    */
   for (i = 0; i < s->b_used; i++)
-    if ((1U << s->log2blkpoolsz) - s->blkpools[i].used >= sizet_length)
+    if (((size_t) 1 << s->log2blkpoolsz) - s->blkpools[i].used >= sizet_length)
       break;
 
   if (i == s->b_used)
